feat(board): add smp support for aarch64 zephyr on hvisor
This commit modifies the CPU configuration for the ROC-RK3588-PC board by reducing the maximum number of CPUs from 4 to 2. Additionally, it updates the device tree to enable UART3 for console and shell communication while disabling UART2. The CPU nodes are also updated to reflect the disabled status for CPU0 and CPU1.

Signed-off-by: Guowei Li <15035660024@163.com>